requirements.txt
streamlit==1.39.0
streamlit-drawable-canvas==0.9.3
Pillow==10.4.0
numpy==1.26.4
